install kernel & python-software-properties separately
The bootstrap script was installing software-properties-common.
The package which contains add-apt-repository is:
python-software-properties

$ apt-file search add-apt-repository
python-software-properties: /usr/bin/add-apt-repository

The kernel failed to install for me in vagrant, so that step should be
done separately to not break the install. The system might already have
a kernel which is fully docker compatible.